Move and rename theming functions to new base_ui folder to speed up startup time
