+++ src/conan_app_launcher/app/base_ui/__init__.py
@@ -1,0 +1,2 @@
+# Basic qt ui functions needed to load before the actual application, to avoid 
+# long import actual uis
