+++ src/conan_app_launcher/ui/dialogs/conan_remove.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 import conan_app_launcher.app as app
-from conan_app_launcher.app.loading import AsyncLoader  # using global module pattern
+from conan_app_launcher.app import AsyncLoader  # using global module pattern
 from conan_app_launcher.app.logger import Logger
 
 from PySide6.QtCore import SignalInstance
