+++ src/conan_app_launcher/ui/views/conan_conf/dialogs/remote_login_dialog.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import List, Optional
 
 import conan_app_launcher.app as app
-from conan_app_launcher.app.loading import AsyncLoader  # using global module pattern
+from conan_app_launcher.app import AsyncLoader  # using global module pattern
 from conan_app_launcher.app.logger import Logger
 from conans.client.cache.remote_registry import Remote
 from PySide6.QtCore import Qt
