+++ src/conan_app_launcher/ui/views/conan_search/model.py
@@ -1,7 +1,7 @@
 from typing import Dict, List, Optional
 
 import conan_app_launcher.app as app
-from conan_app_launcher.app.loading import AsyncLoader  # using global module pattern
+from conan_app_launcher.app import AsyncLoader  # using global module pattern
 from conan_app_launcher.app.logger import Logger
 from conan_app_launcher.conan_wrapper import ConanApi
 from conan_app_launcher.conan_wrapper.types import ConanPkg
